Hasta RF3

+++ src/main/java/uniandes/isis2304/SuperAndes/negocio/Usuario.java
@@ -10,9 +10,9 @@ public class Usuario implements VOUsuario {
 	private String pais;
 	private String ciudad;
 	private String direccion;
-	private int puntos;
-	private long id_TipoUsuario;
-	private long id_Sucursal;
+	private Integer puntos;
+	private Long id_TipoUsuario;
+	private Long id_Sucursal;
 	private long id_Supermercado;
 	
 	public Usuario() {
@@ -26,13 +26,13 @@ public class Usuario implements VOUsuario {
 		this.ciudad = "";
 		this.direccion = "";
 		this.puntos = 0;
-		this.id_TipoUsuario = 0;
-		this.id_Sucursal = 0;
+		this.id_TipoUsuario = (long) 0;
+		this.id_Sucursal = (long) 0;
 		this.id_Supermercado = 0;
 	}
 	
 	public Usuario(long id, long nDocumento, String tipoDocumento, String nombre, String correo,
-			String pais, String ciudad, String direccion, int puntos, long id_TipoUsuario, long id_Sucursal, long id_Supermercado) {
+			String pais, String ciudad, String direccion, Integer puntos, Long id_TipoUsuario, Long id_Sucursal, long id_Supermercado) {
 
 		this.id = id;
 		this.nDocumento = nDocumento;
@@ -116,7 +116,7 @@ public class Usuario implements VOUsuario {
 		return puntos;
 	}
 
-	public void setPuntos(int puntos) {
+	public void setPuntos(Integer puntos) {
 		this.puntos = puntos;
 	}
 	
@@ -124,7 +124,7 @@ public class Usuario implements VOUsuario {
 		return id_TipoUsuario;
 	}
 
-	public void setId_TipoUsuario(long id_TipoUsuario) {
+	public void setId_TipoUsuario(Long id_TipoUsuario) {
 		this.id_TipoUsuario = id_TipoUsuario;
 	}
 
@@ -132,7 +132,7 @@ public class Usuario implements VOUsuario {
 		return id_Sucursal;
 	}
 
-	public void setId_Sucursal(long id_Sucursal) {
+	public void setId_Sucursal(Long id_Sucursal) {
 		this.id_Sucursal = id_Sucursal;
 	}
